+++ Bloch/Cards/OutburstCard.cs
@@ -1,16 +1,13 @@
-﻿using Nanoray.PluginManager;
+﻿using System.Collections.Generic;
+using System.Reflection;
+using Nanoray.PluginManager;
 using Nickel;
 using Shockah.Shared;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 
 namespace Shockah.Bloch;
 
 internal sealed class OutburstCard : Card, IRegisterable
 {
-	private static int GetDataRecursionDepth;
-
 	public static void Register(IPluginPackage<IModManifest> package, IModHelper helper)
 	{
 		helper.Content.Cards.RegisterCard(MethodBase.GetCurrentMethod()!.DeclaringType!.Name, new()
@@ -22,49 +19,49 @@ internal sealed class OutburstCard : Card, IRegisterable
 				rarity = ModEntry.GetCardRarity(MethodBase.GetCurrentMethod()!.DeclaringType!),
 				upgradesTo = [Upgrade.A, Upgrade.B]
 			},
-			Art = helper.Content.Sprites.RegisterSpriteOrDefault(package.PackageRoot.GetRelativeFile("assets/Cards/Outburst.png"), StableSpr.cards_eunice).Sprite,
+			Art = helper.Content.Sprites.RegisterSpriteOrDefault(package.PackageRoot.GetRelativeFile("assets/Cards/Outburst.png"), StableSpr.cards_FumeCannon).Sprite,
 			Name = ModEntry.Instance.AnyLocalizations.Bind(["card", "Outburst", "name"]).Localize
 		});
 	}
 
 	public override CardData GetData(State state)
+		=> upgrade switch
+		{
+			Upgrade.A => new() { cost = 2 },
+			_ => new() { cost = 3 },
+		};
+
+	private bool WillStayInHandOnPlay(State s, Combat c)
 	{
-		GetDataRecursionDepth++;
-		try
-		{
-			return new()
-			{
-				cost = upgrade == Upgrade.A ? 1 : 2,
-				exhaust = upgrade != Upgrade.B,
-				description = ModEntry.Instance.Localizations.Localize(
-					["card", "Outburst", "description", state.route is Combat ? "stateful" : "stateless", upgrade.ToString()],
-					new { Damage = state.route is Combat combat && GetDataRecursionDepth == 1 ? combat.hand.Where(card => card.uuid != uuid).Sum(card => card.GetCurrentCost(state)) : 0 }
-				)
-			};
-		}
-		finally
-		{
-			GetDataRecursionDepth--;
-		}
+		if (!c.hand.Contains(this))
+			return false;
+		if (ModEntry.Instance.Helper.Content.Cards.IsCardTraitActive(s, this, ModEntry.Instance.Helper.Content.Cards.InfiniteCardTrait))
+			return true;
+		if (ModEntry.Instance.Helper.Content.Cards.IsCardTraitActive(s, this, ModEntry.Instance.KokoroApi.Finite.Trait) && ModEntry.Instance.KokoroApi.Finite.GetFiniteUses(s, this) > 1)
+			return true;
+		return false;
 	}
 
 	public override List<CardAction> GetActions(State s, Combat c)
-		=> [new Action { CardId = uuid }];
-
-	private sealed class Action : CardAction
 	{
-		public int? CardId;
-
-		public override void Begin(G g, State s, Combat c)
+		switch (upgrade)
 		{
-			base.Begin(g, s, c);
-
-			var sum = c.hand.Sum(card => card.GetCurrentCost(s));
-			if (CardId is not { } cardId || s.FindCard(cardId) is not { } fromCard)
-				fromCard = null;
-
-			c.QueueImmediate(new AAttack { damage = GetActualDamage(s, sum, card: fromCard) });
-			c.DiscardHand(s);
+			case Upgrade.B:
+			{
+				var extraAmount = c.hand.Count - (WillStayInHandOnPlay(s, c) ? 0 : 1);
+				return
+				[
+					new ADiscard(),
+					new DiscardedVariableHint { ExtraAmount = extraAmount },
+					new AAttack { damage = GetDmg(s, c.DiscardedCardCount + extraAmount), xHint = 1 },
+				];
+			}
+			default:
+				return
+				[
+					new DiscardedVariableHint(),
+					new AAttack { damage = GetDmg(s, c.DiscardedCardCount), xHint = 1 },
+				];
 		}
 	}
 }